refactor: ServiceTest를 서비스 메소드에 적용한다.
