+++ src/test/java/com/wootecam/luckyvickyauction/context/RepositoryTest.java
@@ -13,7 +13,7 @@ import org.springframework.context.annotation.Import;
 // 중간에서 맵핑을 담당하는 CoreRepository 의 경우 자동 생성 대상이 아니므로 직접 명시해 줄 것
 @Import({JpaConfig.class, MemberCoreRepository.class, AuctionCoreRepository.class, ReceiptCoreRepository.class})
 @DataJpaTest
-public class RepositoryTest {
+public abstract class RepositoryTest {
 
     @Autowired
     public AuctionJpaRepository auctionJpaRepository;
