--- src/test/java/com/wootecam/luckyvickyauction/context/SpringBootTest.java
@@ -1,37 +1,0 @@
-package com.wootecam.luckyvickyauction.context;
-
-import static org.springframework.restdocs.mockmvc.MockMvcRestDocumentation.documentationConfiguration;
-import static org.springframework.restdocs.operation.preprocess.Preprocessors.prettyPrint;
-
-import io.restassured.module.mockmvc.RestAssuredMockMvc;
-import io.restassured.module.mockmvc.specification.MockMvcRequestSpecification;
-import org.junit.jupiter.api.BeforeEach;
-import org.junit.jupiter.api.DisplayNameGeneration;
-import org.junit.jupiter.api.DisplayNameGenerator;
-import org.junit.jupiter.api.extension.ExtendWith;
-import org.springframework.boot.test.autoconfigure.web.servlet.AutoConfigureMockMvc;
-import org.springframework.restdocs.RestDocumentationContextProvider;
-import org.springframework.restdocs.RestDocumentationExtension;
-import org.springframework.test.web.servlet.setup.MockMvcBuilders;
-import org.springframework.web.context.WebApplicationContext;
-
-@org.springframework.boot.test.context.SpringBootTest
-@AutoConfigureMockMvc
-@ExtendWith(RestDocumentationExtension.class)
-@DisplayNameGeneration(DisplayNameGenerator.ReplaceUnderscores.class)
-public class SpringBootTest {
-
-    protected MockMvcRequestSpecification docsGiven;
-
-    @BeforeEach
-    void setUp(final WebApplicationContext webApplicationContext,
-               final RestDocumentationContextProvider restDocumentation) {
-        docsGiven = RestAssuredMockMvc.given()
-                .mockMvc(MockMvcBuilders.webAppContextSetup(webApplicationContext)
-                        .apply(documentationConfiguration(restDocumentation)
-                                .operationPreprocessors()
-                                .withRequestDefaults(prettyPrint())
-                                .withResponseDefaults(prettyPrint()))
-                        .build()).log().all();
-    }
-}
