+++ src/test/java/com/wootecam/luckyvickyauction/core/payment/service/PaymentServiceTest.java
@@ -4,50 +4,27 @@ import static org.assertj.core.api.Assertions.assertThat;
 import static org.assertj.core.api.Assertions.assertThatThrownBy;
 import static org.junit.jupiter.api.Assertions.assertAll;
 
+import com.wootecam.luckyvickyauction.context.ServiceTest;
 import com.wootecam.luckyvickyauction.core.auction.domain.Auction;
-import com.wootecam.luckyvickyauction.core.auction.domain.AuctionRepository;
 import com.wootecam.luckyvickyauction.core.auction.domain.ConstantPricePolicy;
 import com.wootecam.luckyvickyauction.core.auction.fixture.AuctionFixture;
-import com.wootecam.luckyvickyauction.core.auction.infra.FakeAuctionRepository;
-import com.wootecam.luckyvickyauction.core.auction.service.AuctionService;
 import com.wootecam.luckyvickyauction.core.member.domain.Member;
-import com.wootecam.luckyvickyauction.core.member.domain.MemberRepository;
 import com.wootecam.luckyvickyauction.core.member.domain.Point;
 import com.wootecam.luckyvickyauction.core.member.domain.Role;
 import com.wootecam.luckyvickyauction.core.member.dto.SignInInfo;
 import com.wootecam.luckyvickyauction.core.member.fixture.MemberFixture;
-import com.wootecam.luckyvickyauction.core.member.infra.FakeMemberRepository;
 import com.wootecam.luckyvickyauction.core.payment.domain.BidHistory;
-import com.wootecam.luckyvickyauction.core.payment.domain.BidHistoryRepository;
 import com.wootecam.luckyvickyauction.core.payment.domain.BidStatus;
-import com.wootecam.luckyvickyauction.core.payment.infra.FakeBidHistoryRepository;
 import com.wootecam.luckyvickyauction.global.exception.BadRequestException;
 import com.wootecam.luckyvickyauction.global.exception.ErrorCode;
 import com.wootecam.luckyvickyauction.global.exception.NotFoundException;
 import com.wootecam.luckyvickyauction.global.exception.UnauthorizedException;
 import java.time.Duration;
 import java.time.ZonedDateTime;
-import org.junit.jupiter.api.BeforeEach;
 import org.junit.jupiter.api.Nested;
 import org.junit.jupiter.api.Test;
 
-// TODO: AuctionService 세부 사항 결정되면 테스트
-class PaymentServiceTest {
-
-    private AuctionRepository auctionRepository;
-    private AuctionService auctionService;
-    private MemberRepository memberRepository;
-    private BidHistoryRepository bidHistoryRepository;
-    private PaymentService paymentService;
-
-    @BeforeEach
-    void setUp() {
-        auctionRepository = new FakeAuctionRepository();
-        auctionService = new AuctionService(auctionRepository);
-        memberRepository = new FakeMemberRepository();
-        bidHistoryRepository = new FakeBidHistoryRepository();
-        paymentService = new PaymentService(auctionService, memberRepository, bidHistoryRepository);
-    }
+class PaymentServiceTest extends ServiceTest {
 
     @Nested
     class process_메소드는 {
@@ -257,11 +234,8 @@ class PaymentServiceTest {
             @Test
             void 환불이_진행된다() {
                 // given
-                Member buyer = MemberFixture.createBuyerWithDefaultPoint();
-                memberRepository.save(buyer);
-
-                Member seller = MemberFixture.createSellerWithDefaultPoint();
-                memberRepository.save(seller);
+                Member buyer = memberRepository.save(MemberFixture.createBuyerWithDefaultPoint());
+                Member seller = memberRepository.save(MemberFixture.createSellerWithDefaultPoint());
 
                 Auction auction = AuctionFixture.createSoldOutAuction();
                 auctionRepository.save(auction);
@@ -299,53 +273,13 @@ class PaymentServiceTest {
         }
 
         @Nested
-        class 만약_환불을_요청한_구매자를_찾을_수_없다면 {
-
-            @Test
-            void 예외가_발생한다() {
-                // given
-                Member buyer = MemberFixture.createBuyerWithDefaultPoint();
-
-                Member seller = MemberFixture.createSellerWithDefaultPoint();
-                memberRepository.save(seller);
-
-                Auction auction = AuctionFixture.createSoldOutAuction();
-                auctionRepository.save(auction);
-
-                ZonedDateTime now = ZonedDateTime.now();
-                BidHistory bidHistory = BidHistory.builder()
-                        .id(1L)
-                        .auctionId(1L)
-                        .productName("test")
-                        .price(100L)
-                        .quantity(1L)
-                        .bidStatus(BidStatus.BID)
-                        .sellerId(seller.getId())
-                        .buyerId(buyer.getId())
-                        .createdAt(now)
-                        .updatedAt(now)
-                        .build();
-                bidHistoryRepository.save(bidHistory);
-
-                // expect
-                assertThatThrownBy(() -> paymentService.refund(new SignInInfo(buyer.getId(), Role.BUYER), 1L))
-                        .isInstanceOf(NotFoundException.class)
-                        .hasMessage("환불할 입찰 내역의 구매자를 찾을 수 없습니다. 구매자 id=1")
-                        .hasFieldOrPropertyWithValue("errorCode", ErrorCode.M002);
-            }
-        }
-
-        @Nested
         class 만약_요청한_사용자가_구매자가_아니라면 {
 
             @Test
             void 예외가_발생한다() {
                 // given
-                Member buyer = MemberFixture.createBuyerWithDefaultPoint();
-                memberRepository.save(buyer);
-
-                Member seller = MemberFixture.createSellerWithDefaultPoint();
-                memberRepository.save(seller);
+                Member buyer = memberRepository.save(MemberFixture.createBuyerWithDefaultPoint());
+                Member seller = memberRepository.save(MemberFixture.createSellerWithDefaultPoint());
 
                 Auction auction = AuctionFixture.createSoldOutAuction();
                 auctionRepository.save(auction);
@@ -380,11 +314,7 @@ class PaymentServiceTest {
             @Test
             void 예외가_발생한다() {
                 // given
-                Member buyer = MemberFixture.createBuyerWithDefaultPoint();
-                memberRepository.save(buyer);
-
-                Member seller = MemberFixture.createSellerWithDefaultPoint();
-                memberRepository.save(seller);
+                Member buyer = memberRepository.save(MemberFixture.createBuyerWithDefaultPoint());
 
                 Auction auction = AuctionFixture.createSoldOutAuction();
                 auctionRepository.save(auction);
@@ -404,11 +334,8 @@ class PaymentServiceTest {
             @Test
             void 예외가_발생한다() {
                 // given
-                Member buyer = MemberFixture.createBuyerWithDefaultPoint();
-                memberRepository.save(buyer);
-
-                Member seller = MemberFixture.createSellerWithDefaultPoint();
-                memberRepository.save(seller);
+                Member buyer = memberRepository.save(MemberFixture.createBuyerWithDefaultPoint());
+                Member seller = memberRepository.save(MemberFixture.createSellerWithDefaultPoint());
 
                 Auction auction = AuctionFixture.createSoldOutAuction();
                 auctionRepository.save(auction);
@@ -443,11 +370,8 @@ class PaymentServiceTest {
             @Test
             void 예외가_발생한다() {
                 // given
-                Member buyer = MemberFixture.createBuyerWithDefaultPoint();
-                memberRepository.save(buyer);
-
-                Member seller = MemberFixture.createSellerWithDefaultPoint();
-                memberRepository.save(seller);
+                Member buyer = memberRepository.save(MemberFixture.createBuyerWithDefaultPoint());
+                Member seller = memberRepository.save(MemberFixture.createSellerWithDefaultPoint());
 
                 Auction auction = AuctionFixture.createSoldOutAuction();
                 auctionRepository.save(auction);
