+++ src/test/java/com/wootecam/luckyvickyauction/global/util/MapperTest.java
@@ -9,7 +9,8 @@ import com.wootecam.luckyvickyauction.core.auction.domain.ConstantPricePolicy;
 import com.wootecam.luckyvickyauction.core.auction.dto.BuyerAuctionSimpleInfo;
 import com.wootecam.luckyvickyauction.core.auction.dto.SellerAuctionSimpleInfo;
 import com.wootecam.luckyvickyauction.core.member.domain.Member;
-import com.wootecam.luckyvickyauction.core.member.fixture.MemberFixture;
+import com.wootecam.luckyvickyauction.core.member.domain.Point;
+import com.wootecam.luckyvickyauction.core.member.domain.Role;
 import com.wootecam.luckyvickyauction.core.payment.domain.BidHistory;
 import com.wootecam.luckyvickyauction.core.payment.dto.BuyerReceiptSimpleInfo;
 import java.time.Duration;
@@ -22,8 +23,20 @@ class MapperTest {
     @Test
     public void 거래내역_entity를_BuyerReceiptSimpleInfo으로_변환하면_도메인의_정보가_동일하게_전달된다() {
         // given
-        Member buyer = MemberFixture.createBuyerWithDefaultPoint();
-        Member seller = MemberFixture.createSellerWithDefaultPoint();
+        Member buyer = Member.builder()
+                .id(1L)
+                .signInId("buyerId")
+                .password("password00")
+                .role(Role.BUYER)
+                .point(new Point(1000L))
+                .build();
+        Member seller = Member.builder()
+                .id(2L)
+                .signInId("sellerId")
+                .password("password00")
+                .role(Role.SELLER)
+                .point(new Point(1000L))
+                .build();
 
         BidHistory history = BidHistory.builder()
                 .id(1L)
